+++ config/boards/corvus.py
@@ -15,7 +15,7 @@ def config(cfg: Config) -> None:
 
     cfg["board"] = {
         "name": "at91sam9g45",
-        "toolchain": "cortexa8hf-neon",
+        "toolchain": "generic-armv7a-hf",
         "power": {
             "on_command": "remote_power at91sam9g45 on",
             "off_command": "remote_power at91sam9g45 off",
@@ -37,7 +37,7 @@ def config(cfg: Config) -> None:
     cfg["uboot"] = (
         {
             "defconfig": "corvus_defconfig",
-            "patchdir": pathlib.PurePosixPath("/home/hws/corvus_patches"),
+            "patchdir": pathlib.PurePosixPath("/home/hws/data/corvus_patches"),
             "test": {
                 "hooks": pathlib.PurePosixPath("/home/hws/hooks/corvus"),
                 "boardname": "corvus",