+++ config/labs/denx.py
@@ -16,7 +16,7 @@ def config(cfg: Config) -> None:
         "keyfile": pathlib.Path.home() / ".ssh" / "id_rsa",
     }
 
-    cfg["tbot.workdir"] = pathlib.PurePosixPath("/work") / username / "tbot2workdir"
+    cfg["tbot.workdir"] = pathlib.PurePosixPath("/work") / username / "tbot-workdir"
 
     cfg["tftp"] = {
         "root": pathlib.PurePosixPath("/tftpboot"),
@@ -29,32 +29,65 @@ def config(cfg: Config) -> None:
         "test.use_venv": False,
     }
 
-    cfg["toolchains"] = {
-        "cortexa8hf-neon": {
-            "path": pathlib.PurePosixPath(
-                "/opt/poky/cortexa8hf-neon-2.4.1/sysroots/x86_64-pokysdk-linux/usr/bin/arm-poky-linux-gnueabi"
-            ),
-            "env_setup_script": pathlib.PurePosixPath(
-                "/opt/poky/cortexa8hf-neon-2.4.1/environment-setup-cortexa8hf-neon-poky-linux-gnueabi"
-            ),
-            "prefix": "arm-poky-linux-gnueabi-",
+    cfg["build"] = {
+        "default": "hercules",
+        "local": "pollux",
+        "pollux": {
+            "ssh_command": f"ssh -i ~/.ssh/id_rsa_nopw {username}@localhost",
+            "scp_command": f"scp -i ~/.ssh/id_rsa_nopw",
+            "scp_address": f"{username}@localhost",
+            "workdir": pathlib.PurePosixPath("/work") / username / "tbot-build",
+            "toolchains": {
+                "cortexa8hf-neon": {
+                    "path": pathlib.PurePosixPath(
+                        "/opt/poky/cortexa8hf-neon-2.4.1/sysroots/x86_64-pokysdk-linux/usr/bin/arm-poky-linux-gnueabi"
+                    ),
+                    "env_setup_script": pathlib.PurePosixPath(
+                        "/opt/poky/cortexa8hf-neon-2.4.1/environment-setup-cortexa8hf-neon-poky-linux-gnueabi"
+                    ),
+                    "prefix": "arm-poky-linux-gnueabi-",
+                },
+                "generic-armv7a-hf": {
+                    "path": pathlib.PurePosixPath(
+                        "/opt/yocto-2.4/generic-armv7a-hf/sysroots/x86_64-pokysdk-linux/usr/bin/arm-poky-linux-gnueabi"
+                    ),
+                    "env_setup_script": pathlib.PurePosixPath(
+                        "/opt/yocto-2.4/generic-armv7a-hf/environment-setup-armv7ahf-neon-poky-linux-gnueabi"
+                    ),
+                    "prefix": "arm-poky-linux-gnueabi-",
+                },
+                "generic-powerpc-e500v2": {
+                    "path": pathlib.PurePosixPath(
+                        "/opt/yocto-2.4/generic-powerpc-e500v2/sysroots/x86_64-pokysdk-linux/usr/bin/powerpc-poky-linux-gnuspe"
+                    ),
+                    "env_setup_script": pathlib.PurePosixPath(
+                        "/opt/yocto-2.4/generic-powerpc-e500v2/environment-setup-ppce500v2-poky-linux-gnuspe"
+                    ),
+                    "prefix": "powerpc-poky-linux-gnuspe-",
+                },
+            },
         },
-        "generic-armv7a-hf": {
-            "path": pathlib.PurePosixPath(
-                "/opt/yocto-2.4/generic-armv7a-hf/sysroots/x86_64-pokysdk-linux/usr/bin/arm-poky-linux-gnueabi"
-            ),
-            "env_setup_script": pathlib.PurePosixPath(
-                "/opt/yocto-2.4/generic-armv7a-hf/environment-setup-armv7ahf-neon-poky-linux-gnueabi"
-            ),
-            "prefix": "arm-poky-linux-gnueabi-",
-        },
-        "generic-powerpc-e500v2": {
-            "path": pathlib.PurePosixPath(
-                "/opt/yocto-2.4/generic-powerpc-e500v2/sysroots/x86_64-pokysdk-linux/usr/bin/powerpc-poky-linux-gnuspe"
-            ),
-            "env_setup_script": pathlib.PurePosixPath(
-                "/opt/yocto-2.4/generic-powerpc-e500v2/environment-setup-ppce500v2-poky-linux-gnuspe"
-            ),
-            "prefix": "powerpc-poky-linux-gnuspe-",
+        "hercules": {
+            "ssh_command": f"ssh -i ~/.ssh/id_rsa_nopw {username}@hercules",
+            "scp_command": f"scp -i ~/.ssh/id_rsa_nopw",
+            "scp_address": f"{username}@hercules",
+            "workdir": pathlib.PurePosixPath("/work") / username / "tbot-build",
+            "toolchains": {
+                "generic-powerpc-e500v2": {
+                    "env_setup_script": pathlib.PurePosixPath(
+                        "/opt/eldk/build/work/hws/p2020/sdk/environment-setup-ppce500v2-poky-linux-gnuspe"
+                    )
+                },
+                "generic-armv7a-hf": {
+                    "env_setup_script": pathlib.PurePosixPath(
+                        "/opt/eldk/build/work/hws/lweimx6/sdk2/environment-setup-armv7ahf-neon-poky-linux-gnueabi"
+                    )
+                },
+                "generic-armv7a": {
+                    "env_setup_script": pathlib.PurePosixPath(
+                        "/opt/eldk/build/work/hws/lweimx6/sdk/environment-setup-armv7a-neon-poky-linux-gnueabi"
+                    )
+                },
+            },
         },
     }