+++ config/labs/local.py
@@ -29,7 +29,7 @@ def config(cfg: Config) -> None:
 
     # Change this to your sdk location
     sdk_base_path = pathlib.Path.home() / "Documents" / "sdk"
-    cfg["toolchains.cortexa8hf-neon"] = {
+    sdk_armv7a_hf = {
         "path": sdk_base_path
         / "sysroots/x86_64-pokysdk-linux"
         / "usr"
@@ -38,4 +38,16 @@ def config(cfg: Config) -> None:
         "env_setup_script": sdk_base_path
         / "environment-setup-cortexa8hf-neon-poky-linux-gnueabi",
         "prefix": "arm-poky-linux-gnueabi-",
+    }
+    cfg["build"] = {
+        "default": "local",
+        "local": {
+            "username": username,
+            "hostname": "localhost",
+            "workdir": cfg["tbot.workdir"] / "build",
+            "toolchains": {
+                "cortexa8hf-neon": sdk_armv7a_hf,
+                "generic-armv7a-hf": sdk_armv7a_hf,
+            },
+        },
     }