+++ tbot/__init__.py
@@ -34,11 +34,17 @@ class TBot:
     """
 
     def __init__(
-        self, config: tbot.config.Config, testcases: dict, new: bool = True
+        self,
+        config: tbot.config.Config,
+        testcases: dict,
+        new: bool = True,
+        interactive: bool = False,
     ) -> None:
         self.config = config
         self.testcases = testcases
         self.layer = 0
+        self.interactive = interactive
+        self._old_inst: typing.Optional[TBot] = None
 
         self.destruct_machines: typing.List[tbot.machine.Machine] = list()
 
@@ -53,8 +59,8 @@ class TBot:
 
     @property
     def shell(self) -> tbot.machine.Machine:
-        """ The default labhost machine """
-        return self.machines["labhost"]
+        """ The default host machine """
+        return self.machines["host"]
 
     @property
     def boardshell(self) -> tbot.machine.MachineBoard:
@@ -90,6 +96,7 @@ class TBot:
         tcs: typing.Union[str, typing.Callable],
         *,
         fail_ok: bool = False,
+        doc: bool = True,
         **kwargs: typing.Any,
     ) -> typing.Any:
         """
@@ -99,6 +106,7 @@ class TBot:
         :type tcs: str or typing.Callable
         :param fail_ok: Whether a failure in this testcase is tolerable
         :type fail_ok: bool
+        :param doc bool: Whether documentation should be generated in this testcase
         :param kwargs: Additional arguments for the testcase
         :type kwargs: dict
         :returns: The return value from the testcase
@@ -107,6 +115,8 @@ class TBot:
         tbot.log_events.testcase_begin(name)
         self.layer += 1
         tbot.log.set_layer(self.layer)
+        previous_doc = tbot.log.LOG_DO_DOC
+        tbot.log.LOG_DO_DOC = previous_doc and doc
         start_time = time.monotonic()
 
         try:
@@ -125,12 +135,14 @@ class TBot:
             run_duration = time.monotonic() - start_time
             tbot.log_events.testcase_end(name, run_duration, False, fail_ok)
             tbot.log.set_layer(self.layer)
+            tbot.log.LOG_DO_DOC = previous_doc
             raise
 
         self.layer -= 1
         run_duration = time.monotonic() - start_time
         tbot.log_events.testcase_end(name, run_duration, True)
         tbot.log.set_layer(self.layer)
+        tbot.log.LOG_DO_DOC = previous_doc
         return retval
 
     def machine(self, mach: tbot.machine.Machine) -> "TBot":
@@ -143,7 +155,9 @@ class TBot:
             statement
         :rtype: TBot
         """
-        new_inst = TBot(self.config, self.testcases, False)
+        new_inst = TBot(
+            self.config, self.testcases, False, interactive=self.interactive
+        )
         new_inst.layer = self.layer
         new_inst.machines = tbot.machine.MachineManager(
             new_inst, self.machines.connection
@@ -162,6 +176,8 @@ class TBot:
         new_inst.machines[mach.unique_machine_name] = new_mach
         if new_mach is not old_mach:
             new_inst.destruct_machines.append(new_mach)
+
+        new_inst._old_inst = self
         return new_inst
 
     def with_board_uboot(self) -> "TBot":
@@ -198,8 +214,15 @@ class TBot:
         for mach in self.destruct_machines:
             # pylint: disable=protected-access
             mach._destruct(self)
+        # Make sure, we don't destruct twice
+        self.destruct_machines = []
 
     def __exit__(
         self, exc_type: typing.Any, exc_value: typing.Any, trceback: typing.Any
     ) -> None:
         self.destruct()
+        # Hack to make this TBot behave like it's parent after the end of a with
+        # statement
+        if self._old_inst is not None:
+            self.machines = self._old_inst.machines
+            self.destruct_machines = self._old_inst.destruct_machines