+++ tbot/builtin/shell_tasks.py
@@ -48,30 +48,58 @@ def setup_tftpdir(
 def cp_to_tftpdir(
     tb: tbot.TBot,
     *,
-    name: typing.Union[str, pathlib.PurePosixPath],
+    source: pathlib.PurePosixPath,
     dest_name: typing.Optional[str] = None,
-    builddir: typing.Optional[pathlib.PurePosixPath] = None,
     tftpdir: tc.TftpDirectory,
 ) -> None:
+    dest_path = tftpdir.path / (source.name if dest_name is None else dest_name)
+
+    tbot.log.debug(f"Copying '{source}' to '{dest_path}'")
+
+    tb.shell.exec0(f"cp {source} {dest_path}")
+
+
+@tbot.testcase
+def retrieve_build_artifact(
+    tb: tbot.TBot,
+    *,
+    buildfile: pathlib.PurePosixPath,
+    buildhost: typing.Optional[str] = None,
+    scp_command: typing.Optional[str] = None,
+    scp_address: typing.Optional[str] = None,
+) -> pathlib.PurePosixPath:
     """
-    Copy a file into the tftp folder
+    Copy artifacts from the buildhost to the labhost
 
-    :param name: Name of the file or path to the file
-    :type name: str or pathlib.PurePosixPath
-    :param dest_name: Name of the file inside the tftp folder, defaults to
-                      the filename of ``name``
-    :type dest_name: str
-    :param builddir: Where to find files if no full path is supplied, defaults to
-                     ``tb.config["uboot.builddir"]``
-    :type builddir: pathlib.PurePosixPath
-    :param tftpdir: Where to put the file
-    :type tftpdir: TftpDirectory
+    :param pathlib.PurePosixPath buildfile: File on the buildhost
+    :param str buildhost: Name of the buildhost if you do not want to use
+                          the default
+    :param str scp_command: SCP command to use for copying (eg ``scp -i <..>``),
+                            defaults to ``tb.config["build.<name>.scp_command"]``
+                            or ``"scp"``
+    :param str scp_address: Address of the form ``<user>@<host>`` of the buildhost,
+                            defaults to ``tb.config["build.<name>.scp_address"]``
+                            or ``"<username>@<hostname>"``
+    :returns: Path where the file has been copied
+    :rtype: pathlib.PurePosixPath
     """
-    builddir = builddir or tb.config["uboot.builddir"]
+    buildhost = buildhost or tb.config["build.default", "<missing>"]
+    bhcfg = f"build.{buildhost}."
+    scp_command = scp_command or tb.config[bhcfg + "scp_command", "scp"]
+    scp_address = scp_address or tb.config[bhcfg + "scp_address", None]
+    scp_address = (
+        scp_address
+        or tb.config[bhcfg + "username"] + "@" + tb.config[bhcfg + "hostname"]
+    )
 
-    source_path = builddir / name if isinstance(name, str) else name
-    dest_path = tftpdir.path / (name if dest_name is None else dest_name)
+    destination = tb.config["tbot.artifactsdir"] / buildfile.name
+    if not isinstance(destination, pathlib.PurePosixPath):
+        raise Exception("config error, tbot.artifactsdir must be a path")
+    tbot.log.debug(f"cp {buildfile} (build) -> {destination} (lab)")
 
-    tbot.log.debug(f"Copying '{source_path}' to '{dest_path}'")
+    tb.machines["labhost-noenv"].exec0(f"mkdir -p {destination.parent}", log_show=False)
+    tb.machines["labhost-noenv"].exec0(
+        f"{scp_command} {scp_address}:{buildfile} {destination}", log_show=False
+    )
 
-    tb.shell.exec0(f"cp {source_path} {dest_path}")
+    return destination