+++ tbot/defaults_config.py
@@ -17,8 +17,11 @@ def config(cfg: Config) -> None:
     ]
 
     # uboot.builddir: U-Boot's repository clone, used to build U-Boot
-    cfg["uboot.builddir"] = cfg[
-        "uboot.builddir", cfg["tbot.workdir"] / f"uboot-{cfg['board.name']}"
+    cfg["uboot.builddir"] = cfg["uboot.builddir", f"uboot-{cfg['board.name']}"]
+
+    # tbot.artifactsdir: Directory where TBot will store artifacts from last build
+    cfg["tbot.artifactsdir"] = cfg[
+        "tbot.artifactsdir", cfg["tbot.workdir"] / f"artifacts-{cfg['board.name']}"
     ]
 
     if cfg["tftp.boarddir", None] is not None: